requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
stripe==12.3.0
Pillow==11.3.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
whitenoise==6.9.0
requests==2.32.4
google-generativeai==0.8.3
adminsortable2==0.15.0